backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
google-genai
python-dotenv
tenacity
pdfplumber
python-multipart
slowapi
sentry-sdk[fastapi]
numpy
